example_skeleton/requirements.txt
pip~=25.2
#constructor_adapter @ git+https://github.com/GiancarloSucci/ConstructorAdapter.git@v0.3.2
#constructor_adapter @ git+ssh://git@github.com/GiancarloSucci/ConstructorAdapter.git@v0.3.2
crewai~=0.117.1
bertopic~=0.17.0
codebleu~=0.7.0
GitPython~=3.1.44
hdbscan~=0.8.40
nltk~=3.9.1
pdfplumber~=0.11.6
Pillow~=11.2.1
python-dotenv~=1.1.0
Requests~=2.32.3
sentence_transformers~=4.1.0
stanza~=1.10.1
tiktoken~=0.9.0
toml~=0.10.2
vermin~=1.6.0
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.9.0
backoff==2.2.1
bcrypt==4.3.0
bertopic==0.17.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.9
ci-info==0.3.0
click==8.1.8
click-plugins==1.1.1.2
codebleu==0.7.0
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.2
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.10
emoji==2.14.1
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
instructor==1.8.2
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.0
json5==0.12.0
json_repair==0.44.1
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langgraph==1.0.0
langchain==1.0.1
langchain-openai==1.0.2
litellm==1.67.2
llvmlite==0.44.0
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
narwhals==1.40.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
openpyxl==3.1.5
openstep_parser==2.0.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pbxproj==4.3.0
pdfminer==20191125
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
plotly==6.1.0
posthog==4.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
prov==2.0.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.3
pynndescent==0.5.13
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyvis==0.3.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==70.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
stanza==1.10.1
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.0
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.51.3
tree-sitter==0.22.3
typer==0.15.4
typing-inspection==0.4.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
uv==0.7.5
uvicorn==0.34.2
uvloop==0.21.0
vermin==1.6.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.41.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
ruff~=0.14.0
jupyterlab==4.4.9
ipywidgets==8.1.7
